# docs/swagger.yaml
basePath: /
definitions:
...
  course.Access:
    properties:
      course_id:
        type: integer
      expires_at:
        type: string
      granted_at:
        type: string
      granted_by:
        type: integer
      id:
        type: integer
      user_id:
        type: integer
    type: object
  course.CreateInput:
    properties:
      description:
        type: string
      is_published:
        type: boolean
      price:
        minimum: 0
        type: number
      slug:
        maxLength: 255
        minLength: 2
        type: string
      title:
        maxLength: 255
        minLength: 2
        type: string
    required:
    - title
    type: object
  course.GrantAccessInput:
    properties:
      expires_at:
        type: string
      user_id:
        type: integer
    required:
    - user_id
    type: object
  course.Model:
    properties:
      author_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_published:
        type: boolean
      price:
        type: number
      slug:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  course.UpdateInput:
    properties:
      description:
        type: string
      is_published:
        type: boolean
      price:
        minimum: 0
        type: number
      slug:
        maxLength: 255
        minLength: 2
        type: string
      title:
        maxLength: 255
        minLength: 2
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: Документация API сервиса it_courses.
  title: API it_courses
  version: "1.0"
paths:
...
  /courses:
    get:
      description: Возвращает список курсов. Для student — только курсы, к которым
        есть доступ
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/course.Model'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Список курсов
      tags:
      - courses
    post:
      consumes:
      - application/json
      description: Создаёт новый курс. Доступно teacher и admin
      parameters:
      - description: Данные курса
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/course.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/course.Model'
        "400":
          description: Bad Request
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
      summary: Создать курс
      tags:
      - courses
...
  /courses/{id}:
    delete:
      description: Удаляет курс по ID. Доступно teacher и admin
      parameters:
      - description: ID курса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Удалить курс
      tags:
      - courses
    put:
      consumes:
      - application/json
      description: Обновляет данные курса по ID. Доступно teacher и admin
      parameters:
      - description: ID курса
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/course.UpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course.Model'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
      summary: Обновить курс
      tags:
      - courses
  /courses/{id}/access:
    get:
      description: Возвращает список пользователей, имеющих доступ к курсу. Только
        admin
      parameters:
      - description: ID курса
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/course.Access'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Список доступов к курсу
      tags:
      - courses
    post:
      consumes:
      - application/json
      description: Предоставляет пользователю доступ к курсу. Только admin
      parameters:
      - description: ID курса
        in: path
        name: id
        required: true
        type: integer
      - description: ID пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/course.GrantAccessInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/course.Access'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
      summary: Выдать доступ к курсу
      tags:
      - courses
  /courses/{id}/access/{userId}:
    delete:
      description: Отзывает доступ пользователя к курсу. Только admin
      parameters:
      - description: ID курса
        in: path
        name: id
        required: true
        type: integer
      - description: ID пользователя
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Отозвать доступ к курсу
      tags:
      - courses
...
  /courses/{slug}:
    get:
      description: Возвращает курс с уроками по slug. Для student проверяется наличие
        доступа
      parameters:
      - description: Slug курса
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course.Model'
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Детали курса
      tags:
      - courses
...
  /my/courses:
    get:
      description: Возвращает список курсов, к которым у текущего пользователя есть
        доступ
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/course.Model'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Мои курсы
      tags:
      - courses
...
swagger: "2.0"